Refactor/strategy authorization #153

+++ src/main/java/com/investmetic/domain/strategy/service/StrategyService.java
@@ -45,14 +45,11 @@ public class StrategyService {
 
 
     @Transactional
-    public void updateVisibility(Long strategyId) {
+    public void updateVisibility(Long strategyId, Long userId) {
         Strategy strategy = strategyRepository.findById(strategyId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.STRATEGY_NOT_FOUND));
 
-        // FIXME : 권한 체크 로직 추가 예정
-//        if (strategy.getCreatedBy() != user) {
-//            throw new BusinessException(ErrorCode.FORBIDDEN_ACCESS);
-//        }
+        verifyUserPermission(strategy, userId);
 
         strategy.setIsPublic(strategy.getIsPublic() == IsPublic.PUBLIC ? IsPublic.PRIVATE : IsPublic.PUBLIC);
     }
@@ -101,14 +98,13 @@ public class StrategyService {
     }
 
     @Transactional
-    public void deleteStrategy(Long strategyId) {
+    public void deleteStrategy(Long strategyId, Long userId) {
         Strategy strategy = strategyRepository.findById(strategyId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.STRATEGY_NOT_FOUND));
 
-        // FIXME : 권한 체크 로직 추가 예정
-//        if (strategy.getCreatedBy() != user) {
-//            throw new BusinessException(ErrorCode.FORBIDDEN_ACCESS);
-//        }
+        verifyUserPermission(strategy, userId);
+
+        s3FileService.deleteFromS3(strategy.getProposalFilePath());
 
         strategyRepository.deleteById(strategyId);
     }
@@ -116,17 +112,12 @@ public class StrategyService {
 
     @Transactional
     public PresignedUrlResponseDto registerStrategy(
-            StrategyRegisterRequestDto requestDto) {
-        // TODO: 추후 삭제 ----------
-        // TODO: 유저 가져오기, tradeType 가져오기, stockType 추가 예정
-        User user = userRepository.findById(1L)
-                .orElseThrow(() -> new BusinessException(ErrorCode.USERS_NOT_FOUND));
-        // 1. TradeType 조회 (예제용 코드로 실제 구현 시 TradeTypeService를 사용하여 조회)
-        TradeType tradeType = tradeTypeRepository.findByTradeTypeId(requestDto.getTradeTypeId())
-                .orElseThrow(() -> new BusinessException(ErrorCode.TRADETYPE_NOT_FOUND));
-        // TODO: 추후 삭제 ----------
+            StrategyRegisterRequestDto requestDto, Long userId) {
+        User user = verifyUser(userId);
 
-        // 2. 제안서 파일 경로 생성 및 Presigned URL 생성
+        TradeType tradeType = tradeTypeRepository.findByTradeTypeIdAndActivateStateTrue(requestDto.getTradeTypeId())
+                .orElseThrow(() -> new BusinessException(ErrorCode.TRADETYPE_NOT_FOUND));
+
         String proposalFilePath = s3FileService.getS3Path(
                 FilePath.STRATEGY_PROPOSAL,
                 requestDto.getProposalFile().getProposalFileName(),
@@ -135,7 +126,6 @@ public class StrategyService {
 
         String presignedUrl = s3FileService.getPreSignedUrl(proposalFilePath);
 
-        // 3. Strategy 생성 및 저장
         Strategy strategy = Strategy.builder()
                 .user(user)
                 .strategyName(requestDto.getStrategyName())
@@ -167,21 +157,14 @@ public class StrategyService {
     @Transactional
     public PresignedUrlResponseDto modifyStrategy(
             Long strategyId,
-            StrategyModifyRequestDto requestDto) {
-        // TODO: 추후 삭제 ----------
-        // TODO: 유저 가져오기, tradeType 가져오기, stockType 추가 예정
-        User user = userRepository.findById(1L)
-                .orElseThrow(() -> new BusinessException(ErrorCode.USER_INFO_NOT_FOUND));
-        // 1. TradeType 조회 (예제용 코드로 실제 구현 시 TradeTypeService를 사용하여 조회)
-        // TODO: 추후 삭제 ----------
-
+            StrategyModifyRequestDto requestDto,
+            Long userId) {
         Strategy strategy = strategyRepository.findById(strategyId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.STRATEGY_NOT_FOUND));
 
-        strategy.modifyStrategy(requestDto.getStrategyName(), requestDto.getDescription());
+        verifyUserPermission(strategy, userId);
 
         if (Boolean.TRUE.equals(requestDto.getProposalModified())) {
-            // 2. 제안서 파일 경로 생성 및 Presigned URL 생성
             String proposalFilePath = s3FileService.getS3Path(
                     FilePath.STRATEGY_PROPOSAL,
                     requestDto.getProposalFile().getProposalFileName(),
@@ -189,8 +172,15 @@ public class StrategyService {
             );
 
             String presignedUrl = s3FileService.getPreSignedUrl(proposalFilePath);
+
+            s3FileService.deleteFromS3(strategy.getProposalFilePath());
+            strategy.modifyStrategyWithProposalFilePath(requestDto.getStrategyName(), requestDto.getDescription(),
+                    proposalFilePath);
+
             return PresignedUrlResponseDto.builder().presignedUrl(presignedUrl).build();
         } else {
+            strategy.modifyStrategyWithoutProposalFilePath(requestDto.getStrategyName(), requestDto.getDescription());
+
             return null;
         }
     }
@@ -202,9 +192,11 @@ public class StrategyService {
         return buildRegisterInfoResponse(tradeTypesDto, stockTypesDto);
     }
 
-    public StrategyModifyInfoResponseDto loadStrategyModifyInfo(Long strategyId) {
+    public StrategyModifyInfoResponseDto loadStrategyModifyInfo(Long strategyId, Long userId) {
         Strategy strategy = strategyRepository.findById(strategyId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.STRATEGY_NOT_FOUND));
+
+        verifyUserPermission(strategy, userId);
 
         TradeTypeDto tradeTypeDto = TradeTypeDto.fromEntity(strategy.getTradeType());
 
@@ -243,5 +235,16 @@ public class StrategyService {
                 .tradeTypes(tradeTypes)
                 .stockTypes(stockTypes)
                 .build();
+    }
+
+    private User verifyUser(Long userId) {
+        return userRepository.findById(userId)
+                .orElseThrow(() -> new BusinessException(ErrorCode.USER_INFO_NOT_FOUND));
+    }
+
+    private void verifyUserPermission(Strategy strategy, Long userId) {
+        if (!strategy.getUser().getUserId().equals(userId)) {
+            throw new BusinessException(ErrorCode.FORBIDDEN_ACCESS);
+        }
     }
 }
